backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
mediapipe
pillow
numpy
scikit-learn
pandas
python-dotenv
requests